更新 ModUpdater.cs

+++ Modules/ModUpdater.cs
@@ -178,6 +178,7 @@ public class ModUpdater
             {
                 if (path.EndsWith(Path.GetFileName(Assembly.GetExecutingAssembly().Location))) continue;
                 if (path.EndsWith("TOHE.dll")) continue;
+                if (!path.EndsWith(".dll")) continue;
                 Logger.Info($"{Path.GetFileName(path)} 已删除", "DeleteOldFiles");
                 File.Delete(path);
             }
